Add BrokerManaging
